Simplify table row drawing
Previously, each view had to calculate widths and such for itself, and
build a format string to go with it.

Instead, it can be simplified down into a series of helper functions
that draw tables of upto 12 columns spread evenly across the terminal
width.

To left align a column, the column span should be negative.