Display generated schedule as root command
Running what-next with no arguments should be the fastest way to show
the schedule.

Some work needs to be put in to handling the case when what-next is run
when a meeting is happening. Should it try to show the amount of time
between meetings? From now until the start of the next one? Some thought
needed.