Hide recurring event definitions
At least in Google Calendar, an instance of a recurring event is
accompanied by a copy of the definition of said event, as two VEVENT
blocks. This was causing what-next to show two meetings happening
simultaneously whenever there was a recurring event, because they both
shared the exact same start and end times.

Showing two meetings instead of one was unexpected behaviour, and a bug.